requirements.txt
anyascii==0.1.7
asgiref==3.3.1
bcp47==0.0.4
beautifulsoup4==4.8.2
certifi==2020.12.5
chardet==4.0.0
Django==3.1.6
django-enumchoicefield==2.0.0
django-filter==2.4.0
django-modelcluster==5.1
django-richtextfield==1.6
django-taggit==1.3.0
django-treebeard==4.4
djangorestframework==3.12.2
draftjs-exporter==2.1.7
et-xmlfile==1.0.1
html5lib==1.1
idna==2.10
jdcal==1.4.1
l18n==2020.6.1
openpyxl==3.0.6
Pillow==8.1.0
pytz==2021.1
requests==2.25.1
six==1.15.0
soupsieve==2.2
sqlparse==0.4.1
tablib==3.0.0
Unidecode==1.2.0
urllib3==1.26.3
wagtail==2.11
wagtailvideos==2.10.2
webencodings==0.5.1
Willow==1.4
xlrd==2.0.1
XlsxWriter==1.3.7
xlwt==1.3.0
